- Исправлены запуски других установщиков provision - Максимальная отмена установок, если уже данный provision установлен

+++ .provision/composer.sh
@@ -1,3 +1,7 @@
 #!/usr/bin/env bash
+if [ -f /bin/composer ]
+	then exit 0
+fi
+
 apt-get install php5-cli -y
 php -r "readfile('https://getcomposer.org/installer');" | php -- --install-dir=/bin --filename=composer
